+++ alinkjs/packages/assetlink/src/pluginModuleLibrary.js
@@ -4,6 +4,7 @@ import * as Quasar from 'quasar';
 import * as AssetlinkPluginApi from 'assetlink-plugin-api';
 import ngeohash from 'ngeohash';
 import haversineDistance from 'haversine-distance';
+import jmespath from 'jmespath';
 
 const libraries = {
   'assetlink-plugin-api': AssetlinkPluginApi,
@@ -13,6 +14,8 @@ const libraries = {
 
   ngeohash,
   'haversine-distance': haversineDistance,
+
+  jmespath,
 
   // TODO: Figure out how to make loading these cleaner/on-demand
   // 'vue-codemirror': import('vue-codemirror'),