+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Make updates non-blocking with the remote when offline
+
 ## [0.10.0] - 2023-02-05
 
 ### Added